requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
sqlalchemy>=2.0.36,<3.0.0
pydantic>=2.10.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
